requirements.txt
Django==1.7.2
django-crispy-forms==1.4.0
markdown==2.5.2
beautifulsoup4==4.3.2